Added Lesson 6
In this lesson, we added a key repeat limiter, added gravity, changed
object movement speed to per second instead of per frame. We added a
variable to the object to make it unable to move, and adjusted the
collision resolution to match. Created a simple world by adding non
moving ground tiles to the bottom of the screen. Fixed a bug that left
"ghost" objects in the gameCell table, since the table needs populated
each frame we just clear the table each frame. I also added a check for
time to live (TTL on the object) to be ignored if it's -1, that way we
can have permanant objects. Also removed the function to remove objects
from the cell table, since it's no longer needed.