Added Lesson 8
Added an attribution file for the art assets, also added a player
character image.

Added a playable character that the camera follow, added the ability to
jump, also made it so the player image flips to face the way the player
is walking.
Also added a variable to track when the objects are resting on top of an
unmoveable object. Found and fixed a bug where I was apply gravity to
the objects position instead of velocity.