Added Lesson 9
This time I've added random level generation, adjusted the player
starting positon to be at the begining of the level(far left). Adjusted
gravity because if felt too floaty, also adjusted the jump height. Fixed
a bug which was incorrectly calculating the offset on drawing objects,
it was using the object min value (upper left), now it uses the objects
position, and set the offset in the draw call.